Rename controller-witness.ts to controller-witness-handler.ts, add witness config:
- Rename server-side handler to controller-witness-handler.ts to
  distinguish from the client-side caller in rep-attestation-frontend
- Add CONTROLLER_WITNESS_CONFIG to attestation-services.ts
- Sync schemas.ts
